requirements.txt
python-dotenv==0.21.1
ibmcloudant>=0.3.3
requests==2.28.2
jinja2==3.1.2
newsapi-python==0.2.6
apache-airflow-providers-postgres